Prevent duplication of keys when calculating normalized values
The normalized values are presented in a dataframe with index being
the year in which the value occurrend. It turns out when determining
the year, some values ended up having the same year several times
which rendered the program broken. We address this issue by ensuring
year is not duplicated