Redefine normalization as x/(max(x) - min(x))
This is somewhat different from the common normalization approach which
is (x - min(x))/(max(x) - min(x)). It is also different from the
normalization approach done when calculatin distance as in distance
method in the pairs trading strategy which is xt/x0.

I do this because I want th emethod to be robust against NaN values in
the dataframe especially when they are the first value